Simplify TunkioFillStrategy & rename it to TunkioFillGenerator

+++ TunkioDLL/Posix/TunkioPosixFile.cpp
@@ -82,9 +82,9 @@ namespace Tunkio
 		return m_size;
 	}
 
-	std::pair<bool, uint64_t> File::Write(const uint8_t* data, const uint32_t size) const
+	std::pair<bool, uint64_t> File::Write(const uint8_t* data, const uint64_t size) const
 	{
-		ssize_t result = write(m_fileDescriptor, data, size);
+		ssize_t result = write(m_fileDescriptor, data, static_cast<size_t>(size));
 
 		if (result == -1)
 		{
